+++ src/commands/mod.rs
@@ -15,5 +15,6 @@ pub mod send;
 pub mod signal;
 pub mod status;
 pub mod ui;
+pub mod uninstall;
 pub mod view;
 pub mod watch;